# refactor: use expand directories from tinyglobby