--- a/src/utils.ts
+++ b/src/utils.ts
@@ -57,78 +57,38 @@ type GenerateFileListOptions = {
 
 export async function generateFileList({
   rootDir,
-  include = ["*"],
+  include = ["**"],
   exclude,
   excludeFn,
 }: GenerateFileListOptions): Promise<string[]> {
-  const firstPass = await glob(include, {
+  const paths = await glob(include, {
     cwd: rootDir,
     ignore: exclude,
-    onlyFiles: false,
+    expandDirectories: true,
   });
 
-  const files = new Set<string>();
-  const dirs = new Set<string>();
-  for (const path of firstPass) {
-    if (path.endsWith("/")) {
-      dirs.add(`${path}**`);
-    } else {
-      files.add(path);
-    }
-  }
-
-  const secondPass = await glob(Array.from(dirs), {
-    cwd: rootDir,
-    ignore: exclude,
-  });
-  for (const path of secondPass) {
-    files.add(path);
-  }
-
-  let result = Array.from(files);
+  let result = Array.from(paths);
   if (excludeFn) {
     result = result.filter((path) => !excludeFn(path));
   }
-
-  result.sort();
   return result;
 }
 
 export function generateFileListSync({
   rootDir,
-  include = ["*"],
+  include = ["**"],
   exclude,
   excludeFn,
 }: GenerateFileListOptions): string[] {
-  const firstPass = globSync(include, {
+  const paths = globSync(include, {
     cwd: rootDir,
     ignore: exclude,
-    onlyFiles: false,
+    expandDirectories: true,
   });
 
-  const files = new Set<string>();
-  const dirs = new Set<string>();
-  for (const path of firstPass) {
-    if (path.endsWith("/")) {
-      dirs.add(`${path}**`);
-    } else {
-      files.add(path);
-    }
-  }
-
-  const secondPass = globSync(Array.from(dirs), {
-    cwd: rootDir,
-    ignore: exclude,
-  });
-  for (const path of secondPass) {
-    files.add(path);
-  }
-
-  let result = Array.from(files);
+  let result = Array.from(paths);
   if (excludeFn) {
     result = result.filter((path) => !excludeFn(path));
   }
-
-  result.sort();
   return result;
 }