… bun.lock
… package.json