## What does this PR do?

## Relevant implementation details

## How did you test this?


<!-- This is an auto-generated comment: release notes by coderabbit.ai -->
## Summary by CodeRabbit

- New Features
  - Recursive file matching enabled by default to capture nested files.
  - CLI now computes and times fingerprints, persists baseline/current snapshots, and shows diffs when they differ.

- Refactor
  - File discovery streamlined to a single-pass for faster, more consistent results; excludes applied after discovery and output follows glob order.

- Benchmarks
  - Added Expensify benchmark tasks to the fingerprint suite and updated benchmark repo list.

- Chores
  - Ignored fingerprint artifacts and added dev dependencies for diff tooling.
<!-- end of auto-generated comment: release notes by coderabbit.ai -->