@mdjastrzebski 2025-09-21
📊 Performance Comparison (vs baseline):
| Task name                     | Baseline latency (ms) | Current latency (ms) | Change (ms)      |
| ----------------------------- | --------------------- | -------------------- | ---------------- |
| lodash-sync                   | 16.95                 | 16.24                | -0.71 (-4.2%)    |
| lodash                        | 19.24                 | 18.23                | -1.01 (-5.3%)    |
| express-sync                  | 8.50                  | 7.90                 | -0.60 (-7.1%)    |
| express                       | 8.58                  | 8.04                 | -0.53 (-6.2%)    |
| react-sync                    | 322.20                | 316.85               | -5.35 (-1.7%)    |
| react                         | 374.77                | 358.17               | -16.60 (-4.4%)   |
| react-native-sync             | 699.63                | 494.52               | -205.11 (-29.3%) |
| react-native                  | 660.36                | 513.71               | -146.65 (-22.2%) |
| react-native (null hash)-sync | 225.11                | 66.72                | -158.40 (-70.4%) |
| react-native (null hash)      | 135.40                | 25.05                | -110.35 (-81.5%) |